requirements.txt
python = "3.11.4"
langchain = "^0.2.11"
langchain-openai = "^0.1.20"
langchain-community = "^0.2.10"
langchain-chroma = "^0.1.2"
pypdf = "^4.3.1"
jupyterlab = "^4.2.4"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"

